chore(#1323): merge 4 thin async wrappers into sync counterparts
These files were pure asyncio.to_thread() forwarding with zero unique
logic. Merging them eliminates 4 files and ~590 lines of indirection.

- async_manager.py → manager.py (AsyncReBACManager)
- async_namespace_manager.py → namespace_manager.py (AsyncNamespaceManager)
- async_router.py → router.py (AsyncVFSRouter)
- async_agent_registry.py → agent_registry.py (AsyncAgentRegistry)

All importers updated (src/ + tests/). NOT merging async_bridge.py,
async_permissions.py, or async_files.py which have unique logic.
