+++ src/nexus/bricks/rebac/async_bridge.py
@@ -28,7 +28,7 @@ from concurrent.futures import Future
 from typing import TYPE_CHECKING, Any
 
 if TYPE_CHECKING:
-    from nexus.bricks.rebac.async_manager import AsyncReBACManager
+    from nexus.bricks.rebac.manager import AsyncReBACManager
 
 logger = logging.getLogger(__name__)
 
@@ -158,8 +158,7 @@ class AsyncReBACBridge:
         Issue #621: Engine creation delegates to RecordStoreABC instead of
         calling create_engine() directly.
         """
-        from nexus.bricks.rebac.async_manager import AsyncReBACManager
-        from nexus.bricks.rebac.manager import ReBACManager
+        from nexus.bricks.rebac.manager import AsyncReBACManager, ReBACManager
 
         if self._injected_engine is not None:
             sync_engine = self._injected_engine
