+++ src/nexus/system_services/agents/__init__.py
@@ -3,9 +3,8 @@
 Canonical location for agent registry, provisioning, and lifecycle.
 """
 
-from nexus.system_services.agents.agent_registry import AgentRegistry
+from nexus.system_services.agents.agent_registry import AgentRegistry, AsyncAgentRegistry
 from nexus.system_services.agents.agent_service import AgentService
-from nexus.system_services.agents.async_agent_registry import AsyncAgentRegistry
 
 __all__ = [
     "AgentRegistry",
