+++ tests/unit/services/test_async_agent_registry.py
@@ -8,7 +8,7 @@ import pytest
 
 from nexus.contracts.agent_types import AgentRecord, AgentState
 from nexus.contracts.protocols.agent_registry import AgentInfo, AgentRegistryProtocol
-from nexus.system_services.agents.async_agent_registry import AsyncAgentRegistry, _to_agent_info
+from nexus.system_services.agents.agent_registry import AsyncAgentRegistry, _to_agent_info
 from tests.unit.core.protocols.test_conformance import assert_protocol_conformance
 
 # ---------------------------------------------------------------------------
