+++ tests/unit/services/test_async_namespace_manager.py
@@ -4,11 +4,12 @@ from unittest.mock import MagicMock
 
 import pytest
 
-from nexus.bricks.rebac.async_namespace_manager import (
+from nexus.bricks.rebac.namespace_manager import (
     AsyncNamespaceManager,
+    MountEntry,
+    NamespaceMount,
     _to_namespace_mount,
 )
-from nexus.bricks.rebac.namespace_manager import MountEntry, NamespaceMount
 from nexus.contracts.protocols.namespace_manager import NamespaceManagerProtocol
 from tests.unit.core.protocols.test_conformance import assert_protocol_conformance
